requirements.txt
flask==3.0.0
openai==0.28
python-dotenv==1.0.0
SpeechRecognition==3.10.0
pyttsx3==2.90
pyaudio==0.2.13
Werkzeug==3.0.1